feat: refactor editor functionality to an editor store
