+++ apps/web/app/space/(entity)/[id]/[entityId]/default-entity-page.tsx
@@ -5,6 +5,7 @@ import { Suspense } from 'react';
 
 import { AppConfig, Environment } from '~/core/environment';
 import { API, Subgraph } from '~/core/io';
+import { EditorProvider } from '~/core/state/editor-store';
 import { EntityStoreProvider } from '~/core/state/entity-page-store/entity-store-provider';
 import { MoveEntityProvider } from '~/core/state/move-entity-store';
 import { DEFAULT_PAGE_SIZE } from '~/core/state/triple-store/constants';
@@ -55,24 +56,29 @@ export default async function DefaultEntityPage({ params, searchParams }: Props)
         spaceId={props.spaceId}
         initialTriples={props.triples}
         initialSchemaTriples={[]}
-        initialBlockIdsTriple={props.blockIdsTriple}
-        initialBlockTriples={props.blockTriples}
       >
-        <MoveEntityProvider>
-          <EntityPageCover avatarUrl={avatarUrl} coverUrl={coverUrl} />
-          <EntityPageContentContainer>
-            <EditableHeading spaceId={props.spaceId} entityId={props.id} name={props.name} triples={props.triples} />
-            <EntityPageMetadataHeader id={props.id} spaceId={props.spaceId} types={types} />
-            <Spacer height={40} />
-            <Editor shouldHandleOwnSpacing />
-            <ToggleEntityPage {...props} filterId={filterId} filterValue={filterValue} typeId={typeId} />
-            <Spacer height={40} />
-            <Suspense fallback={<EntityReferencedByLoading />}>
-              <EntityReferencedByServerContainer entityId={props.id} name={props.name} spaceId={params.id} />
-            </Suspense>
-          </EntityPageContentContainer>
-          <MoveEntityReview />
-        </MoveEntityProvider>
+        <EditorProvider
+          id={props.id}
+          spaceId={props.spaceId}
+          initialBlockIdsTriple={props.blockIdsTriple}
+          initialBlockTriples={props.blockTriples}
+        >
+          <MoveEntityProvider>
+            <EntityPageCover avatarUrl={avatarUrl} coverUrl={coverUrl} />
+            <EntityPageContentContainer>
+              <EditableHeading spaceId={props.spaceId} entityId={props.id} name={props.name} triples={props.triples} />
+              <EntityPageMetadataHeader id={props.id} spaceId={props.spaceId} types={types} />
+              <Spacer height={40} />
+              <Editor shouldHandleOwnSpacing />
+              <ToggleEntityPage {...props} filterId={filterId} filterValue={filterValue} typeId={typeId} />
+              <Spacer height={40} />
+              <Suspense fallback={<EntityReferencedByLoading />}>
+                <EntityReferencedByServerContainer entityId={props.id} name={props.name} spaceId={params.id} />
+              </Suspense>
+            </EntityPageContentContainer>
+            <MoveEntityReview />
+          </MoveEntityProvider>
+        </EditorProvider>
       </EntityStoreProvider>
     </TypesStoreServerContainer>
   );
